pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "async_llm"
version = "0.0.1"
description = "A Python package for asynchronous LLM integration."
readme = "README.md"
authors = [{ name = "Shamal De Silva", email = "shamalgithub@gmail.com" }, {name = "Sahan Ruwantha", email="sahanr.silva@proton.me"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "fastapi>=0.112.2",
    "httpx>=0.27.2",
    "pydantic>=2.8.2",
    "pytest>=8.3.2",
    "python-dotenv>=1.0.1",
    "setuptools>=58.1.0",
    "pipreqs>=0.5.0",
    "httpx[http2]>=0.27.2",
    "pytest-asyncio>=0.24.0"
]
